+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -8,10 +8,13 @@
 
 #include <media_kit_libs_windows_video/media_kit_libs_windows_video_plugin_c_api.h>
 #include <media_kit_video/media_kit_video_plugin_c_api.h>
+#include <video_player_custom/video_player_custom_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
   MediaKitLibsWindowsVideoPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("MediaKitLibsWindowsVideoPluginCApi"));
   MediaKitVideoPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("MediaKitVideoPluginCApi"));
+  VideoPlayerCustomPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("VideoPlayerCustomPluginCApi"));
 }