+++ example/windows/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   media_kit_libs_windows_video
   media_kit_video
+  video_player_custom
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST